+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Upgraded
+- Calypsonet Terminal Reader API `1.3.0` -> Keypop Reader API `2.0.0`
+- Calypsonet Terminal Calypso API `1.8.0` -> Keypop Calypso Card API `2.0.0`
+- Keyple Service Library `2.3.1` -> `3.0.0`
+- Keyple Calypso Card Library `2.3.5` -> `3.0.0`
+- Keyple Util Library `2.3.0` -> `2.3.1`
+
+### Added
+New dependencies
+- Keypop Crypto Legacy SAM API `0.3.0`
+- Keyple Calypso Crypto LegacySAM Library `0.4.0`
 
 ## [2023.06.01]
 ### Upgraded